% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des2beta.R
\name{des2beta}
\alias{des2beta}
\title{Estimate skewness and kurtosis based on truncated normal distribution in a two group sample}
\usage{
des2beta(
  vmean = vmean,
  vsd = vsd,
  lo = lo,
  hi = hi,
  design = "ind",
  cor_xy = NULL,
  showFigure = FALSE,
  ...
)
}
\arguments{
\item{vmean}{a vector of two sample means of the beta distributed data}

\item{vsd}{a vector of two sample standard deviation of the beta distributed data}

\item{lo}{a vector of two minimum possible value}

\item{hi}{a vector of two maximum possible value}

\item{design}{if design = "ind", the two groups are independent; if design = "dep", the two groups are dependent}

\item{cor_xy}{if design = "dep", the correlation between the two groups needs to be provided for further analysis.}

\item{showFigure}{when showFigure = TRUE, it will display the plots with theoretical normal and beta curves.}

\item{...}{other arguments}
}
\description{
This function can be used to calculate the skewness and kurtosis based on the truncated normal distribution.
Also, this function estimate the mean and standard deviation of the parent distribution.
}
\examples{
data('beta_mdat')
ex <- beta_mdat[2, ]
#independent two groups
des2beta(vmean=c(ex$m1, ex$m2), vsd=c(ex$sd1, ex$sd2),
hi = c(ex$hi1, ex$hi2), lo = c(ex$lo1, ex$lo2), design = "ind")

# dependent two groups with correlation equal to .3
ex <- beta_mdat[12, ]
des2beta(vmean=c(ex$m1, ex$m2), vsd=c(ex$sd1, ex$sd2),
hi = c(ex$hi1, ex$hi2), lo = c(ex$lo1, ex$lo2), design = "dep", cor_xy = .3)
}
\references{
\insertRef{johnson1995continuous}{detectnorm}

\insertRef{smithson2006better}{detectnorm}

\insertRef{olkin2014cons}{detectnorm}
}
\seealso{
\code{\link{desbeta}}
}
