% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des2trunc.R
\name{des2trunc}
\alias{des2trunc}
\title{Estimate skewness and kurtosis based on truncated normal distribution in two groups in one study}
\usage{
des2trunc(vmean = vmean, vsd = vsd, lo = lo, hi = hi, showFigure = FALSE, ...)
}
\arguments{
\item{vmean}{sample mean of the beta distributed data}

\item{vsd}{sample standard deviation of the beta distributed data}

\item{lo}{minimum possible value}

\item{hi}{maximum possible value}

\item{showFigure}{when showFigure = TRUE, it will display the plots with theoretical normal and beta curves.}

\item{...}{other arguments}
}
\description{
This function can be used to calculate the skewness and kurtosis based on the truncated normal distribution.
Also, this function estimate the mean and standard deviation of the parent distribution.
}
\examples{
data('trun_mdat')
ex <- trun_mdat[4, ]
#independent two groups
des2trunc(vmean=c(ex$m1, ex$m2), vsd=c(ex$sd1, ex$sd2),
hi = c(ex$hi1, ex$hi2), lo = c(ex$lo1, ex$lo2))

}
\seealso{
\code{\link{desbeta}}
}
