% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desbeta.R
\name{desbeta}
\alias{desbeta}
\title{Estimate skewness and kurtosis based on Beta distribution in one sample}
\usage{
desbeta(vmean, vsd, lo, hi, showFigure = FALSE, ...)
}
\arguments{
\item{vmean}{sample mean of the beta distributed data}

\item{vsd}{sample standard deviation of the beta distributed data}

\item{lo}{minimum possible value}

\item{hi}{maximum possible value}

\item{showFigure}{when showFigure = TRUE, it will display the plots with theoretical normal and beta curves.}

\item{...}{other arguments}
}
\description{
This function can be used to calculate the skewness and kurtosis based on the Beta distribution. Also, this function estimate the shape parameters alpha and beta.
}
\examples{
data('beta_mdat')
desbeta(vmean=beta_mdat$m2[6], vsd=beta_mdat$sd2[6],
hi = beta_mdat$hi2[6], lo = beta_mdat$lo2[6], showFigure = TRUE)

}
\references{
\insertRef{johnson1995continuous}{detectnorm}

\insertRef{smithson2006better}{detectnorm}

\insertRef{olkin2014cons}{detectnorm}
}
\seealso{
\code{\link{destrunc}}
}
