% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectnorm.R
\name{detectnorm}
\alias{detectnorm}
\title{Calculate skewness and kurtosis based on Beta or truncated normal distribution in a meta-analysis for SMD (Two independent groups)}
\usage{
detectnorm(
  m1i,
  sd1i,
  n1i,
  lo1i,
  hi1i,
  m2i,
  sd2i,
  n2i,
  lo2i,
  hi2i,
  data,
  showFigure = FALSE,
  distri = "beta",
  ...
)
}
\arguments{
\item{m1i}{vector to the means of first group}

\item{sd1i}{vector to specifiy the standard deviation of first group}

\item{n1i}{vector to specify the sample size of first group}

\item{lo1i}{vector to specify the possible minimum of the first group}

\item{hi1i}{vector to specify the possible maximum of the first group}

\item{m2i}{vector to the means of second group}

\item{sd2i}{vector to specifiy the standard deviation of second group}

\item{n2i}{vector to specify the sample size of second group}

\item{lo2i}{vector to specify the possible minimum of the second group}

\item{hi2i}{vector to specify the possible maximum of the second group}

\item{showFigure}{when showFigure = TRUE, it will display all the plots (within the result as a list, result$fig) with theoretical normal curve and the truncated normal curve.}

\item{...}{other arguments}
}
\description{
This function can be used to calculate the skewness and kurtosis based on the Beta distribution with the dataset used to conduct meta-analysis.
}
\examples{
\dontrun{
data("metadat")
ex <- detectnorm(m1i = m1,sd1i = sd1,n1i = n1,
hi1i = p.max,lo1i = 0,m2i = m2,sd2i = sd2,n2i = n2,
hi2i = p.max,lo2i=0,distri = "beta", data = metadat)
}

}
