% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2norm.R
\encoding{UTF-8}
\name{meta2norm}
\alias{meta2norm}
\title{Calculate skewness and kurtosis based on Beta distribution in a meta-analysis for SMD (Two independent or dependent groups)}
\usage{
meta2norm(
  m1,
  m2,
  sd1,
  sd2,
  hi1,
  hi2,
  lo1,
  lo2,
  cor_xy = NA,
  distri = "beta",
  design = "ind",
  metadat = NULL,
  showFigure = FALSE,
  ...
)
}
\arguments{
\item{m1}{vector to the means of first group}

\item{m2}{vector to the means of second group}

\item{sd1}{vector to specifiy the standard deviation of first group}

\item{sd2}{vector to specifiy the standard deviation of second group}

\item{hi1}{vector to specify the possible maximum of the first group}

\item{hi2}{vector to specify the possible maximum of the second group}

\item{lo1}{vector to specify the possible minimum of the first group}

\item{lo2}{vector to specify the possible minimum of the second group}

\item{cor_xy}{specify the correlations between the two groups when they are dependent.}

\item{distri}{Beta distribution is used when using `distri = "beta"`; Truncated normal distribution is used when using `distri = "truncnorm"`}

\item{design}{specify if the two groups are independent or dependent.}

\item{metadat}{The data.frame of meta-analytic data with means, stantand deviations, minimums, maximums, design (specify whether "dep" or "ind"), and correlations (for dependent two groups).}

\item{showFigure}{when showFigure = TRUE, it will display all the plots (within the result as a list, result$fig) with theoretical normal curve and the truncated normal curve.}

\item{...}{other arguments}
}
\value{
The output of the data frame adding some columns of the possible skewness and kurtosis for each groups.
}
\description{
This function can be used to calculate the skewness and kurtosis based on the Beta distribution with the dataset used to conduct meta-analysis.
}
\examples{
#truncated normal data
data("trun_mdat")
ex <- meta2norm(m1 = m1,sd1 = sd1, hi1 = hi1, lo1 = lo1,
m2 = m2, sd2 = sd2, hi2 = hi2, lo2 = lo2, distri = "trunc",
metadat = trun_mdat)
head(ex)
#extremely non-normal data
data("beta_mdat")
ex2 <- meta2norm(m1 = m1,sd1 = sd1, hi1 = hi1, lo1 = lo1,
m2 = m2, sd2 = sd2, hi2 = hi2, lo2 = lo2, distri = "beta",
metadat = beta_mdat)
head(ex2)
mean(ex2$skew1)#sample skewness calculated from the sample
mean(ex2$g1skew) #estimated using beta
}
\references{
\insertRef{barr1999mean}{detectnorm}

\insertRef{johnson1995continuous}{detectnorm}

\insertRef{robert1995simulation}{detectnorm}

\insertRef{shah1966estimation}{detectnorm}

\insertRef{smithson2006better}{detectnorm}

\insertRef{sun2020influence}{detectnorm}
}
