% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadat.R
\docType{data}
\name{metadat}
\alias{metadat}
\title{Example meta-analysis}
\format{
A data frame with 38 rows and 19 variables:
\describe{
  \item{study}{the authors and year for the study included in the meta-analysis}
  \item{n1}{sample size of patients with schizophrenia}
  \item{m1}{mean of patients with schizophrenia}
  \item{sd1}{standard deviation of patients with schizophrenia}
  \item{cv1}{coefficient of variability of patients with schizophrenia}
  \item{n2}{sample size of healthy control subjects}
  \item{m2}{mean of healthy control subjects}
  \item{sd2}{standard deviation of healthy control subjects}
  \item{cv2}{coefficient of variability of healthy control subjects}
  \item{p.max}{possible maximum of the items}
  \item{d}{standardized mean difference}
}
}
\usage{
metadat
}
\description{
This dataset contains the individual studies in the meta-analysis
by Bora et al (2009). They meta-analyzed 38 studies about theory of mind impairement in schizophrenia.
}
\keyword{datasets}
