% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnonnorm.R
\name{rnonnorm}
\alias{rnonnorm}
\title{Non-normal Distribution
Generating Non-normal data with specified skewness and kurtosis using Fleishman's Method}
\usage{
rnonnorm(n, mean = 0, sd = 1, skew = 0, kurt = 0)
}
\arguments{
\item{n}{snumber of observations}

\item{mean}{mean}

\item{sd}{standard deviation}

\item{skew}{skewness}

\item{kurt}{kurtosis}
}
\description{
This function can be used to generate non-normal data with specified skewness and kurtosis using Fleishman's Power Method.
n, mean = 0, sd = 1, skew = 0, kurt = 0
}
