% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desbeta.R
\name{desbeta}
\alias{desbeta}
\title{Calculate skewness and kurtosis based on Beta distribution in one group}
\usage{
desbeta(vmean, vsd, lo = 0, hi = 1, showFigure = FALSE, ...)
}
\arguments{
\item{vmean}{sample mean of the truncated data}

\item{vsd}{sample standard deviation of the truncated data}

\item{lo}{minimum possible value}

\item{hi}{maximum possible value}

\item{showFigure}{when showFigure = TRUE, it will display the plots with theoretical normal curve and the truncated normal curve.}

\item{...}{other arguments}
}
\description{
This function can be used to calculate the skewness and kurtosis based on the Beta distribution. Also, this function estimate the shape parameters alpha and beta.
}
\examples{
data("metadat")
desbeta(vmean=dat$m2[6], vsd=dat$sd2[6],
hi = dat$p.max[6],showFigure = T)
}
\seealso{
\code{\link{destrunc}}
}
