% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destrunc.R
\name{destrunc}
\alias{destrunc}
\title{Calculate skewness and kurtosis based on truncated normal distribution in one group}
\usage{
destrunc(vmean, vsd, lo = 0, hi, showFigure = FALSE, ...)
}
\arguments{
\item{vmean}{sample mean of the truncated data}

\item{vsd}{sample standard deviation of the truncated data}

\item{lo}{minimum possible value}

\item{hi}{maximum possible value}

\item{showFigure}{when showFigure = TRUE, it will display the plots with theoretical normal curve and the truncated normal curve.}

\item{...}{other arguments}
}
\description{
This function can be used to calculate the skewness and kurtosis based on the truncated normal distribution. Also, it can be used to estimate the mean and variance of the parent distribution (the distribution before truncated).
}
\examples{
data("metadat")
destrunc(vmean=dat$m2[6], vsd=dat$sd2[6],
hi = dat$p.max[6],showFigure = T)
}
\seealso{
\code{\link{desbeta}}
}
